find id and lost btn close
