+++ src/components/NoticeCategoryItem/NoticeCategoryItem.jsx
@@ -1,5 +1,7 @@
 import { useState } from "react";
 import { useDeleteNoticeMutation } from "redux/noticesApi";
+import Modal from "components/Modal/Modal";
+import ModalNotice from "components/ModalNotice";
 import {
   Item,
   ImageThumb,
@@ -16,6 +18,19 @@ import unlike from "img/unlike.svg";
 const NoticeCategoryItem = ({ id, name, title, birthday, breed, male, location, price, image, onModalOpen }) => {
   const [isFavorite, setFavorite] = useState(false);
   const [deleteNotice, { isLoading: isDeleting }] = useDeleteNoticeMutation();
+  const [isOpenModalNotice, setIsOpenModalNotice] = useState(false);
+
+  const openModalNotice = e => {
+    if (e) {
+      setIsOpenModalNotice(true);
+    }
+  };
+
+  const closeModalNotice = e => {
+    if (e) {
+      setIsOpenModalNotice(false);
+    }
+  };
 
   const currentAge = date => {
     let today = new Date();
@@ -39,50 +54,58 @@ const NoticeCategoryItem = ({ id, name, title, birthday, breed, male, location, 
   };
 
   return (
-    <Item>
-      <ImageThumb>
-        <Image src="https://cdn.pixabay.com/photo/2021/10/27/19/09/cat-6748193_960_720.jpg" alt={title}></Image>
-        <Category>Sell</Category>
-        <BtnFavorite type="button" onClick={() => console.log("isFavorite")}>
-          <img src={unlike} alt="unlike" />
-        </BtnFavorite>
-      </ImageThumb>
-      <div>
-        <Title>{title}</Title>
-        <ContainerDescription>
-          <Table>
-            <tbody>
-              <tr>
-                <td>Name:</td>
-                <td>{name}</td>
-              </tr>
-              <tr>
-                <td>Breed:</td>
-                <td>{breed}</td>
-              </tr>
-              <tr>
-                <td>Place:</td>
-                <td>{location}</td>
-              </tr>
-              <tr>
-                <td>Age:</td>
-                <td>{currentAge(birthday)}</td>
-              </tr>
-              <tr>
-                <td>Price:</td>
-                <td>{price}</td>
-              </tr>
-            </tbody>
-          </Table>
-        </ContainerDescription>
-      </div>
-      <ButtonMore active="true" type="button" onClick={() => onModalOpen(id)}>
-        Learn more
-      </ButtonMore>
-      <button type="button" disabled={isDeleting} onClick={() => deleteNotice(id)}>
-        Delete
-      </button>
-    </Item>
+    <>
+      {isOpenModalNotice && (
+        <Modal onClose={closeModalNotice}>
+          <ModalNotice onClose={closeModalNotice} id={id} />
+        </Modal>
+      )}
+      *
+      <Item>
+        <ImageThumb>
+          <Image src="https://cdn.pixabay.com/photo/2021/10/27/19/09/cat-6748193_960_720.jpg" alt={title}></Image>
+          <Category>Sell</Category>
+          <BtnFavorite type="button" onClick={() => console.log("isFavorite")}>
+            <img src={unlike} alt="unlike" />
+          </BtnFavorite>
+        </ImageThumb>
+        <div>
+          <Title>{title}</Title>
+          <ContainerDescription>
+            <Table>
+              <tbody>
+                <tr>
+                  <td>Name:</td>
+                  <td>{name}</td>
+                </tr>
+                <tr>
+                  <td>Breed:</td>
+                  <td>{breed}</td>
+                </tr>
+                <tr>
+                  <td>Place:</td>
+                  <td>{location}</td>
+                </tr>
+                <tr>
+                  <td>Age:</td>
+                  <td>{currentAge(birthday)}</td>
+                </tr>
+                <tr>
+                  <td>Price:</td>
+                  <td>{price}</td>
+                </tr>
+              </tbody>
+            </Table>
+          </ContainerDescription>
+        </div>
+        <ButtonMore active="true" type="button" onClick={() => openModalNotice(id)}>
+          Learn more
+        </ButtonMore>
+        <button type="button" disabled={isDeleting} onClick={() => deleteNotice(id)}>
+          Delete
+        </button>
+      </Item>
+    </>
   );
 };
